backend/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.41.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
pooch==1.8.1
protobuf==5.26.1
PyMatting==1.1.12
PyWavelets==1.4.1
referencing==0.35.1
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.1
scikit-image==0.21.0
scipy==1.10.1
sympy==1.12
tifffile==2023.7.10
tqdm==4.66.4
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.1
